Portland, Oregon Weather Data [SNWSDPDX]

A shortened daily dataset for Portland, Oregon containing the daily maximum, minimum temperatures in degrees Fahrenheit and precipitation and snow amounts. The variables are as follows [as obtained from the national weather service]:

SNWSDPDX

Format

A data frame with 1095 rows and 8 variables:

Day

an integer Day of the Month (1--31)

YR

four digit calendar year (2017--2019)

MO

two digit calendar month (1--12)

TX

maximum average temperature F (39--87.5)

TN

maximum average temperature F (27.97--61.4)

PR

Precipitation in in (0--15.24)

SN

Snow in in (0--19)

date

a date [Date format] (2005 Jan 1--2019 Dec 31)